Move common gauge stuff to gauge.py
